1、大改动

+++ apps/chipin/src/main/java/com/tinx/java/chipin/entity/query/UserLotteryQuery.java
@@ -26,11 +26,11 @@ public class UserLotteryQuery extends ChipinQuery {
     /**
      * 唯一标识
      */
-    private Integer id;
+    private Long id;
     /**
      * 用户ID
      */
-    private Integer userId;
+    private Long userId;
     /**
      * 网站ID
      */
